Allow fingerprint duration more than ~9 hours
Fixes https://github.com/acoustid/acoustid-server/issues/43

+++ acoustid/tables.py
@@ -310,7 +310,7 @@ fingerprint = Table(
     metadata,
     Column("id", Integer, primary_key=True),
     Column("fingerprint", ARRAY(Integer), nullable=False),
-    Column("length", SmallInteger, CheckConstraint("length>0"), nullable=False),
+    Column("length", Integer, CheckConstraint("length>0"), nullable=False),
     Column("bitrate", SmallInteger, CheckConstraint("bitrate>0")),
     Column("format_id", Integer),
     Column(
